move file_utilities to cudf_test
